Move loaders/schema/test files to loaders/test
