+++ packages/db/src/loaders/test/compilationSources/MagicSquare.sol
