+++ packages/db/src/loaders/test/compilationSources/VyperStorage.vy
